fix freeze on shutdown
closing the sdl audio subsystem would fail since the callback was
still waiting for a lock.

this bug was introduced in 0fda42d20818d2e2ff0261a91ba73bc148bba13f
"patch by deathpudding", the latest commit done by pokemonhacker.

that patch introduced 3 different locks, which seems over-engineered
and it's possible that it is the reason for the stuttering sound
i experience, and maybe even for the recurring hangs (every ~15 sec,
the emulator hangs for like half a second).

usually when i see a comment in code like "patch by randomgamer"
the alarmbell rings, since most of them are idiots... and especially
those that spray their nickname all over the place...
so it may well be that reverting that patch fixes several issues.